+++ CompuCell3D/core/Demos/SBMLSolverExamples/SBMLSolverAntimony/SBMLSolverAntimony/Simulation/SBMLSolverAntimony.py
@@ -1,0 +1,10 @@
+from cc3d import CompuCellSetup
+from SBMLSolverAntimonySteppables import SBMLSolverSteppable
+from SBMLSolverAntimonySteppables import IdFieldVisualizationSteppable
+
+
+CompuCellSetup.register_steppable(steppable=SBMLSolverSteppable(frequency=1))
+CompuCellSetup.register_steppable(steppable=IdFieldVisualizationSteppable(frequency=1))
+
+
+CompuCellSetup.run()