+++ CompuCell3D/core/Demos/SBMLSolverExamples/SBMLSolverAntimony/SBMLSolverAntimony2/Simulation/SBMLSolverAntimony2.py
@@ -1,0 +1,13 @@
+from cc3d import CompuCellSetup
+from SBMLSolverAntimony2Steppables import SBMLSolverSteppable
+from SBMLSolverAntimony2Steppables import IdFieldVisualizationSteppable
+from SBMLSolverAntimony2Steppables import SecretionSteppable
+
+
+CompuCellSetup.register_steppable(steppable=SBMLSolverSteppable(frequency=1))
+
+CompuCellSetup.register_steppable(steppable=IdFieldVisualizationSteppable(frequency=1))
+
+CompuCellSetup.register_steppable(steppable=SecretionSteppable(frequency=1))
+
+CompuCellSetup.run()